requirements.txt
python-telegram-bot
flask
flask-restx
sqlalchemy
cryptography
python-dotenv
reportlab
pandas
werkzeug
